+++ controller/v1/vo/network.go
@@ -1,0 +1,9 @@
+package vo
+
+func CreateNetworkDisconnectVO(errs []error) []string {
+	results := make([]string, len(errs))
+	for i, err := range errs {
+		results[i] = err.Error()
+	}
+	return results
+}