+++ pkg/mageddo/uuid/uuid.go
@@ -3,6 +3,7 @@ package uuid
 import (
 "crypto/rand"
 "fmt"
+	"strings"
 )
 
 func UUID() (uuid string) {
@@ -12,5 +13,9 @@ func UUID() (uuid string) {
 		fmt.Println("Error: ", err)
 		return
 	}
-	return fmt.Sprintf("%X-%X-%X-%X-%X", b[0:4], b[4:6], b[6:8], b[8:10], b[10:])
+	return fmt.Sprintf("%x-%x-%x-%x-%x", b[0:4], b[4:6], b[6:8], b[8:10], b[10:])
+}
+
+func TruncatedUUID(size int) string {
+	return strings.Replace(UUID(), "-", "", -1)[0:size]
 }